requirements.txt
langchain
python-dotenv
huggingface_hub
openai